# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task API",
    "version": "1.0.0",
    "description": "A CRUD API for managing a to-do list, secured with Supabase Auth."
  },
  "servers": [
    { "url": "http://localhost:3000" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/": {
      "get": {
        "summary": "API info",
        "description": "Returns basic info about the API.",
        "responses": {
          "200": { "description": "API info returned" }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health check",
        "description": "Confirms the server is running.",
        "responses": {
          "200": { "description": "Server is healthy" }
        }
      }
    },
    "/public/info": {
      "get": {
        "summary": "Public info",
        "description": "Returns a public welcome message. No authentication required.",
        "responses": {
          "200": { "description": "Public message returned" }
        }
      }
    },
    "/auth/signup": {
      "post": {
        "summary": "Sign up",
        "description": "Creates a new user account via Supabase Auth.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "example": "test@example.com" },
                  "password": { "type": "string", "example": "password123" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User created" },
          "400": { "description": "Missing or invalid email/password" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Log in",
        "description": "Authenticates a user and returns an access token and refresh token.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "example": "test@example.com" },
                  "password": { "type": "string", "example": "password123" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login successful, tokens returned" },
          "400": { "description": "Missing email or password" },
          "401": { "description": "Invalid login credentials" }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Log out",
        "description": "Ends the current user's session. Requires a valid bearer token.",
        "security": [
          { "bearerAuth": [] }
        ],
        "responses": {
          "204": { "description": "Logged out successfully" },
          "401": { "description": "Missing, malformed, or invalid/expired token" }
        }
      }
    },
    "/protected/profile": {
      "get": {
        "summary": "Get profile",
        "description": "Returns the authenticated user's id, email, and account creation date. Requires a valid bearer token.",
        "security": [
          { "bearerAuth": [] }
        ],
        "responses": {
          "200": { "description": "User profile returned" },
          "401": { "description": "Missing, malformed, or invalid/expired token" }
        }
      }
    },
    "/protected/dashboard": {
      "get": {
        "summary": "Get dashboard",
        "description": "Returns a welcome message for the authenticated user. Requires a valid bearer token.",
        "security": [
          { "bearerAuth": [] }
        ],
        "responses": {
          "200": { "description": "Welcome message returned" },
          "401": { "description": "Missing, malformed, or invalid/expired token" }
        }
      }
    },
    "/tasks": {
      "get": {
        "summary": "List all tasks",
        "description": "Returns the full list of tasks.",
        "responses": {
          "200": { "description": "List of tasks returned" }
        }
      },
      "post": {
        "summary": "Create a task",
        "description": "Creates a new task. Requires a non-empty title.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string", "example": "Buy milk" }
                },
                "required": ["title"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Task created" },
          "400": { "description": "Missing or empty title" }
        }
      }
    },
    "/tasks/{id}": {
      "get": {
        "summary": "Get a single task",
        "description": "Returns one task by id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Task returned" },
          "404": { "description": "Task not found" }
        }
      },
      "put": {
        "summary": "Update a task",
        "description": "Updates a task's title and/or done status.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string", "example": "Buy bread" },
                  "done": { "type": "boolean", "example": true }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Task updated" },
          "400": { "description": "Invalid body" },
          "404": { "description": "Task not found" }
        }
      },
      "delete": {
        "summary": "Delete a task",
        "description": "Removes a task by id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "204": { "description": "Task deleted" },
          "404": { "description": "Task not found" }
        }
      }
    }
  }
}